Dev: unittest: Adjust unit test code for qdevice.py
